fix for vercel verify-voucher error -> Error: EROFS: read-only file system, mkdir '/var/task/uploads'
