+++ api/index.js
@@ -1,2 +1,2 @@
-const app = require('../server/app');
+const app = require('../server/index');
 module.exports = app;
